Floating y

+++ bnn-predictive/preds/svgp.py
@@ -94,7 +94,7 @@ class SVGPNTK():
             K_t = torch.transpose(K, dim0=1, dim1=0)
             lambda_inv = lambdas_class**(-1) #*torch.eye(self.x.shape[0])
             A = K @ (lambda_inv * torch.eye(self.x.shape[0]) )@ K_t
-            alpha_f[i_class] = torch.matmul(K, self.y) #torch.linalg.solve(A, y.float())
+            alpha_f[i_class] = torch.matmul(K, self.y.float()) #torch.linalg.solve(A, y.float())
             beta_f[i_class] = A #torch.linalg.solve(lambda_inv * torch.eye(gram.shape[0]) + K, torch.eye(K.shape[0]))
         return alpha_f, beta_f
